PrimaryNav-related file renaming and import fixes
