+++ packages/manager/src/components/PrimaryNav/PrimaryNav.styles.ts
@@ -25,12 +25,7 @@ const useStyles = makeStyles((theme: Theme) => ({
     position: 'relative',
     display: 'flex',
     alignItems: 'center',
-    padding: `
-        ${theme.spacing(2) - 2}px
-        0
-        ${theme.spacing(1) + theme.spacing(1) / 2}px
-        ${theme.spacing(4)}px
-      `,
+    padding: '14px 14px 12px',
     '& svg': {
       maxWidth: theme.spacing(3) + 91
     }
@@ -40,54 +35,51 @@ const useStyles = makeStyles((theme: Theme) => ({
       opacity: 0
     }
   },
+  linkGroup: {
+    marginBottom: theme.spacing(3)
+  },
   listItem: {
     display: 'flex',
     alignItems: 'center',
     position: 'relative',
     cursor: 'pointer',
-    maxHeight: 42,
+    height: 36,
+    lineHeight: 0,
     transition: theme.transitions.create(['background-color']),
-    padding: `
-        ${theme.spacing(1.5)}px
-        ${theme.spacing(4) - 2}px
-        ${theme.spacing(1.5) - 1}px
-        ${theme.spacing(4) + 1}px
-      `,
-    '&:hover, &:focus': {
-      backgroundColor: theme.bg.primaryNavActiveBG,
+    '& p': {
+      marginTop: 0,
+      marginBottom: 0
+    },
+    padding: '12px 20px',
+    '&:focus': {
+      textDecoration: 'none'
+    },
+    '&:hover': {
       textDecoration: 'none',
+      border: 'red',
+      backgroundImage: 'linear-gradient(98deg, #38584B 1%, #3A5049 166%)',
       '& $linkItem': {
         color: 'white'
       },
       '& svg': {
-        fill: 'white',
-        '& *': {
-          stroke: 'white'
-        }
+        color: theme.color.greenCyan,
+        fill: theme.color.greenCyan
       }
     },
     '& .icon': {
+      color: '#CFD0D2',
       marginRight: theme.spacing(2),
       '& svg': {
-        width: 26,
-        height: 26,
-        transform: 'scale(1.75)',
-        fill: theme.color.primaryNavText,
-        transition: theme.transitions.create(['fill']),
-        '&.small': {
-          transform: 'scale(1)'
-        },
+        display: 'flex',
+        alignItems: 'center',
+        width: 20,
+        height: 20,
         '&:not(.wBorder) circle, & .circle': {
           display: 'none'
-        },
-        '& *': {
-          transition: theme.transitions.create(['stroke']),
-          stroke: theme.color.primaryNavText
         }
       }
     }
   },
-  listItemCollapsed: {},
   collapsible: {
     fontSize: '0.9rem'
   },
@@ -102,68 +94,18 @@ const useStyles = makeStyles((theme: Theme) => ({
     }
   },
   active: {
-    backgroundColor: theme.bg.primaryNavActiveBG,
-    '&:before': {
-      content: "''",
-      borderStyle: 'solid',
-      borderWidth: `
-          ${theme.spacing(1) + 11}px
-          ${theme.spacing(1) + 6}px
-          ${theme.spacing(1) + 11}px
-          0
-        `,
-      borderColor: `transparent ${theme.bg.primaryNavActive} transparent transparent`,
-      position: 'absolute',
-      right: 0,
-      top: theme.spacing(1) === 8 ? 2 : 0
+    backgroundImage: 'linear-gradient(98deg, #38584B 1%, #3A5049 166%)',
+    textDecoration: 'none',
+    '& svg': {
+      color: theme.color.greenCyan
     },
-    '&:hover': {
-      textDecoration: 'none',
-      '&:before': {
-        content: "''",
-        borderColor: `transparent ${theme.bg.primaryNavActive} transparent transparent`
-      }
-    },
-    [theme.breakpoints.down('sm')]: {
-      '&:before': {
-        display: 'none'
-      }
-    },
-    '&.listItemCollapsed': {
-      '&:before': {
-        top: theme.spacing(1) === 8 ? 2 : 4
-      }
-    }
+    '&:hover': {}
   },
   spacer: {
     padding: 25
   },
   divider: {
     backgroundColor: 'rgba(0, 0, 0, 0.12)'
-  },
-  settings: {
-    width: 30,
-    margin: `auto auto 16px`,
-    alignItems: 'center',
-    justifyContent: 'center',
-    display: 'flex',
-    color: '#e7e7e7',
-    transition: theme.transitions.create(['color']),
-    '& svg': {
-      transition: theme.transitions.create(['transform'])
-    },
-    '&:hover': {
-      color: theme.color.green
-    }
-  },
-  settingsCollapsed: {
-    margin: `auto 16px 16px ${theme.spacing(4) - 1}px`
-  },
-  activeSettings: {
-    color: theme.color.green,
-    '& svg': {
-      transform: 'rotate(90deg)'
-    }
   },
   menu: {},
   paper: {
